Refactor: PerpetualMintStorage risks (#53)
* refactor: normalize PerpetualMintStorage risks as uint64 data types

* refactor: consolidate totalTokenRisk & tokenRisk

+++ test/facets/L2/AssetHandler/behaviors/withdrawERC721Assets.t.sol
@@ -233,7 +233,7 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint256 tokenRisk = uint256(
+        uint64 tokenRisk = uint64(
             uint256(vm.load(address(this), tokenRiskStorageSlot))
         );
 
@@ -268,8 +268,8 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint256 totalDepositorRisk = uint256(
-            vm.load(address(this), totalDepositorRiskStorageSlot)
+        uint64 totalDepositorRisk = uint64(
+            uint256(vm.load(address(this), totalDepositorRiskStorageSlot))
         );
 
         // this assertion proves that the total risk for the depositor in the collection was decremented correctly
@@ -283,7 +283,7 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint128 totalRisk = uint128(
+        uint64 totalRisk = uint64(
             uint256(vm.load(address(this), totalRiskStorageSlot))
         );
